Moved some stuff out of concepts.hpp to avoid circular dependencies.
ssize needs range_difference_t which is in concepts.hpp but concepts.hpp include primitives.hpp where ssize is.
